+++ backend/src/migrations.rs
@@ -292,15 +292,16 @@ pub enum SchemaVersion {
     AlterPltTokensAddCurrentSupplyColumn,
     #[display("0046: Alter transaction table to add sponsored transaction fields")]
     AlterTxnAddSponsoredTxn,
-    #[display("0047: Add partial index for nonzero PLT holders by token")]
+    #[display("0047: Plt Accounts statement table to track balance changes")]
+    PltAccountsStatements,
+    #[display("0048: Add partial index for nonzero PLT holders by token")]
     IndexPltHolderNonZero,
 }
 impl SchemaVersion {
     /// The minimum supported database schema version for the API.
     /// Fails at startup if any breaking (destructive) database schema versions
     /// have been introduced since this version.
-    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion =
-        SchemaVersion::AlterPltTokensAddCurrentSupplyColumn;
+    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion = SchemaVersion::IndexPltHolderNonZero;
     /// The latest known version of the schema.
     const LATEST: SchemaVersion = SchemaVersion::IndexPltHolderNonZero;
 
@@ -369,6 +370,7 @@ impl SchemaVersion {
             SchemaVersion::AddTokenModulePauseUnpauseStatus => false,
             SchemaVersion::NormalizePltCumulativeTransferAmount => false,
             SchemaVersion::AlterPltTokensAddCurrentSupplyColumn => false,
+            SchemaVersion::PltAccountsStatements => false,
             SchemaVersion::AlterTxnAddSponsoredTxn => false,
             SchemaVersion::IndexPltHolderNonZero => false,
         }
@@ -426,6 +428,7 @@ impl SchemaVersion {
             SchemaVersion::AddTokenModulePauseUnpauseStatus => false,
             SchemaVersion::NormalizePltCumulativeTransferAmount => false,
             SchemaVersion::AlterPltTokensAddCurrentSupplyColumn => false,
+            SchemaVersion::PltAccountsStatements => false,
             SchemaVersion::AlterTxnAddSponsoredTxn => false,
             SchemaVersion::IndexPltHolderNonZero => false,
         }
@@ -753,11 +756,20 @@ impl SchemaVersion {
             SchemaVersion::AlterTxnAddSponsoredTxn => {
                 tx.as_mut()
                     .execute(sqlx::raw_sql(include_str!(
-                        "./migrations/m0047_index_plt_holder_non_zero.sql"
+                        "./migrations/m0047-plt-accounts-statement.sql"
+                    )))
+                    .await?;
+                SchemaVersion::PltAccountsStatements
+            }
+            SchemaVersion::PltAccountsStatements => {
+                tx.as_mut()
+                    .execute(sqlx::raw_sql(include_str!(
+                        "./migrations/m0048_index_plt_holder_non_zero.sql"
                     )))
                     .await?;
                 SchemaVersion::IndexPltHolderNonZero
             }
+            
             SchemaVersion::IndexPltHolderNonZero => unimplemented!(
                 "No migration implemented for database schema version {}",
                 self.as_i64()