+++ frontend/src/components/StableCoin/Holders.vue
@@ -43,6 +43,11 @@
 							</TableRow>
 						</TableBody>
 					</Table>
+					<LoadMore
+						v-if="pltHolderData?.pltAccountsByTokenId?.pageInfo"
+						:page-info="pltHolderData.pltAccountsByTokenId.pageInfo"
+						:on-load-more="loadMore"
+					/>
 				</CarouselSlide>
 			</FtbCarousel>
 		</div>
@@ -53,6 +58,7 @@ import FtbCarousel from '~/components/molecules/FtbCarousel.vue'
 import CarouselSlide from '~/components/molecules/CarouselSlide.vue'
 import BWCubeLogoIcon from '~/components/icons/BWCubeLogoIcon.vue'
 import Filter from '~/components/StableCoin/Filter.vue'
+import LoadMore from '~/components/LoadMore.vue'
 import { usePltTokenHolderQuery } from '~/queries/usePltTokenHolderQuery'
 import type { PltAccountAmount } from '~/types/generated'
 import { usePagedData } from '~/composables/usePagedData'
@@ -75,25 +81,39 @@ const lastNTransactions = ref(TransactionFilterOption.Top20)
 const coinId = props.coinId
 const totalSupply = props.totalSupply
 
-const { pagedData, addPagedData } = usePagedData<PltAccountAmount>(
-	[],
-	lastNTransactions.value,
-	lastNTransactions.value
-)
-
-const queryFirst = ref(lastNTransactions.value)
+const pageSize = ref(lastNTransactions.value)
+const { first, last, after, before, pagedData, addPagedData, loadMore } =
+	usePagedData<PltAccountAmount>([], pageSize.value, lastNTransactions.value)
 
 watch(
 	() => lastNTransactions.value,
 	newValue => {
-		queryFirst.value = newValue
+		// Update page size and reset pagination when filter changes
+		pageSize.value = newValue
+		first.value = newValue
+		after.value = undefined
+		before.value = undefined
+		last.value = undefined
 		pagedData.value = []
 	}
 )
 
+// Update first/last before query to ensure correct page size
+watch([first, last], ([firstVal, lastVal]) => {
+	if (firstVal) {
+		first.value = pageSize.value
+	}
+	if (lastVal) {
+		last.value = pageSize.value
+	}
+})
+
 const { data: pltHolderData, fetching: pltHolderLoading } =
 	usePltTokenHolderQuery(coinId, {
-		first: queryFirst,
+		first,
+		last,
+		after,
+		before,
 	})
 
 watch(